+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/viewmodel/messages/MessageListViewModelFactory.kt
@@ -48,6 +48,8 @@ import io.getstream.chat.android.ui.common.utils.AttachmentConstants
  * @param dateSeparatorHandler Handler that determines when the date separators should be visible.
  * @param threadDateSeparatorHandler Handler that determines when the thread date separators should be visible.
  * @param messagePositionHandler Determines the position of the message inside a group.
+ * @param showDateSeparatorInEmptyThread Configures if we show a thread separator when threads are empty or not. Adds the
+ * separator item when the value is `true`.
  *
  * @see MessageListHeaderViewModel
  * @see MessageListViewModel
@@ -70,6 +72,7 @@ public class MessageListViewModelFactory @JvmOverloads constructor(
     private val threadDateSeparatorHandler: DateSeparatorHandler =
         DateSeparatorHandler.getDefaultThreadDateSeparatorHandler(),
     private val messagePositionHandler: MessagePositionHandler = MessagePositionHandler.defaultHandler(),
+    private val showDateSeparatorInEmptyThread: Boolean = false,
 ) : ViewModelProvider.Factory {
 
     private val factories: Map<Class<*>, () -> ViewModel> = mapOf(
@@ -99,6 +102,7 @@ public class MessageListViewModelFactory @JvmOverloads constructor(
                     dateSeparatorHandler = dateSeparatorHandler,
                     threadDateSeparatorHandler = threadDateSeparatorHandler,
                     messagePositionHandler = messagePositionHandler,
+                    showDateSeparatorInEmptyThread = showDateSeparatorInEmptyThread,
                 ),
                 chatClient = chatClient
             )